recover ability to do case-insensitive equality restriction on TextAttribute
Note that this is implemented to always use Pattern, but that's OK, because
it's easy for an implementation of Data to determine that the whole pattern
has no wildcards, and then use = instead of like in that case.

An alternative approach would be to introduce TextValue and a TextRange
interface abstracting over Pattern and TextValue. This might be better and
more flexible, if slightly more complicated.